Fix Exercise Index Bug with random indexes
Due to update order weirdness, pattern exercises added by key would all
have the same index and be deleted together. To avoid this, I switched
to random indexes instead of sequential. This also revealed a but where
the inded update was the only thing causing the re-render of the
exercise list, so i added a blank useEffect so the component gets
re-rendered whenever exercise configs changes. ugh.

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>